Move Scenario to Kite Extras
