+++ KITE-Janus-Test/src/main/java/io/cosmosoftware/kite/janus/steps/ScreenshotStep.java
@@ -1,6 +1,6 @@
 package io.cosmosoftware.kite.janus.steps;
 
-import io.cosmosoftware.kite.janus.Scenario;
+import io.cosmosoftware.kite.instrumentation.Scenario;
 import io.cosmosoftware.kite.report.Reporter;
 import io.cosmosoftware.kite.steps.TestStep;
 import org.openqa.selenium.WebDriver;
